fix: migration foreign key 欄位使用相同欄位結構
例如：bike_models table 的 bike_brand 欄位，比照 bike_brands table 的 id 欄位使用 unsignedBigInteger
另外把原本寫 ->nullable(false) 這個多餘程式碼刪除

+++ database/migrations/2023_05_23_080104_create_parts_issues_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
     {
         Schema::create('parts_issues', function (Blueprint $table) {
             $table->id();
-            $table->integer('bike_parts_id');
+            $table->unsignedBigInteger('bike_parts_id');
             $table->string('error_code')->unique();
             $table->string('description');
             $table->timestamps();
